yaml reflection generator separated from bare.cc
