+++ src/shdc/main.cc
@@ -105,6 +105,14 @@ int main(int argc, const char** argv) {
         case format_t::BARE:
             output_err = bare_t::gen(args, inp, spirvcross, bytecode);
             break;
+        case format_t::BARE_YAML:
+            output_err = bare_t::gen(args, inp, spirvcross, bytecode);
+            if (output_err.valid)
+                break;
+
+            output_err = yaml_t::gen(args, inp, spirvcross, bytecode);
+            break;
+
         case format_t::SOKOL_ZIG:
             output_err = sokolzig_t::gen(args, inp, spirvcross, bytecode);
             break;
